Add tailwind to project
I have not spent the time to get really good with css. Heard tailwind is
a good approach. The utility first approach is something I am skeptical
of because I have been spoiled by good designers providing the necessary
css / html. This type of work is a grey area between the work of a
designer and an engineer but I lean towards it being in better hands
with a designer because they can easily identify where the css misses
the designs.