Update terraform with imported manually created resources
I was playing around a lot with aws and afterword wanted to put them
down in terraform. Luckily terraform has an import function for this.
This is an exceptionally simple setup. Free ec2 instance with a docker
conatiner hosting a static html (docker container spun up manually).
vpc, subnet, load balancer, route 53 all set up just for the static
html. Not sure the lb is necessary.