Updates to how redux fetches & general interface
I don't really like writing these commits on a personal education repo
because most of them are like 'played around with this thing more'.
This usage of redux is definitely wrong, there are several bugs with
getting the wrong data and the 'useEffect if not pending' etc is turning
out to be a terrible way to prevent continuous fetching. I really need
to find a way to embed this logic in a dynamic way into either a
selector or the async thunk for all json api calls.